Migrated zigzag indicator to the class pattern.
