+++ js/Stock/Indicators/ZIGZAG/ZIGZAGIndicator.js
@@ -8,8 +8,28 @@
  *
  * */
 'use strict';
-import BaseSeries from '../../Core/Series/Series.js';
-// im port './SMAIndicator.js';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../../Core/Series/Series.js';
+var SMAIndicator = BaseSeries.seriesTypes.sma;
+import U from '../../../Core/Utilities.js';
+var merge = U.merge, extend = U.extend;
+/* *
+ *
+ * Class
+ *
+ * */
 /**
  * The Zig Zag series type.
  *
@@ -19,60 +39,26 @@ import BaseSeries from '../../Core/Series/Series.js';
  *
  * @augments Highcharts.Series
  */
-BaseSeries.seriesType('zigzag', 'sma', 
-/**
- * Zig Zag indicator.
- *
- * This series requires `linkedTo` option to be set.
- *
- * @sample stock/indicators/zigzag
- *         Zig Zag indicator
- *
- * @extends      plotOptions.sma
- * @since        6.0.0
- * @product      highstock
- * @requires     stock/indicators/indicators
- * @requires     stock/indicators/zigzag
- * @optionparent plotOptions.zigzag
- */
-{
-    /**
-     * @excluding index, period
-     */
-    params: {
-        /**
-         * The point index which indicator calculations will base - low
-         * value.
+var ZIGZAGIndicator = /** @class */ (function (_super) {
+    __extends(ZIGZAGIndicator, _super);
+    function ZIGZAGIndicator() {
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        /* *
          *
-         * For example using OHLC data, index=2 means the indicator will be
-         * calculated using Low values.
-         */
-        lowIndex: 2,
-        /**
-         * The point index which indicator calculations will base - high
-         * value.
+         *  Properties
          *
-         * For example using OHLC data, index=1 means the indicator will be
-         * calculated using High values.
-         */
-        highIndex: 1,
-        /**
-         * The threshold for the value change.
-         *
-         * For example deviation=1 means the indicator will ignore all price
-         * movements less than 1%.
-         */
-        deviation: 1
+         * */
+        _this.data = void 0;
+        _this.points = void 0;
+        _this.options = void 0;
+        return _this;
     }
-}, 
-/**
- * @lends Highcharts.Series#
- */
-{
-    nameComponents: ['deviation'],
-    nameSuffixes: ['%'],
-    nameBase: 'Zig Zag',
-    getValues: function (series, params) {
+    /* *
+     *
+     *  Functions
+     *
+     * */
+    ZIGZAGIndicator.prototype.getValues = function (series, params) {
         var lowIndex = params.lowIndex, highIndex = params.highIndex, deviation = params.deviation / 100, deviations = {
             'low': 1 + deviation,
             'high': 1 - deviation
@@ -164,8 +150,66 @@ BaseSeries.seriesType('zigzag', 'sma',
             xData: xData,
             yData: yData
         };
-    }
+    };
+    /**
+     * Zig Zag indicator.
+     *
+     * This series requires `linkedTo` option to be set.
+     *
+     * @sample stock/indicators/zigzag
+     *         Zig Zag indicator
+     *
+     * @extends      plotOptions.sma
+     * @since        6.0.0
+     * @product      highstock
+     * @requires     stock/indicators/indicators
+     * @requires     stock/indicators/zigzag
+     * @optionparent plotOptions.zigzag
+     */
+    ZIGZAGIndicator.defaultOptions = merge(SMAIndicator.defaultOptions, {
+        /**
+         * @excluding index, period
+         */
+        params: {
+            /**
+             * The point index which indicator calculations will base - low
+             * value.
+             *
+             * For example using OHLC data, index=2 means the indicator will be
+             * calculated using Low values.
+             */
+            lowIndex: 2,
+            /**
+             * The point index which indicator calculations will base - high
+             * value.
+             *
+             * For example using OHLC data, index=1 means the indicator will be
+             * calculated using High values.
+             */
+            highIndex: 1,
+            /**
+             * The threshold for the value change.
+             *
+             * For example deviation=1 means the indicator will ignore all price
+             * movements less than 1%.
+             */
+            deviation: 1
+        }
+    });
+    return ZIGZAGIndicator;
+}(SMAIndicator));
+extend(ZIGZAGIndicator.prototype, {
+    nameComponents: ['deviation'],
+    nameSuffixes: ['%'],
+    nameBase: 'Zig Zag'
 });
+BaseSeries.registerSeriesType('zigzag', ZIGZAGIndicator);
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default ZIGZAGIndicator;
 /**
  * A `Zig Zag` series. If the [type](#series.zigzag.type) option is not
  * specified, it is inherited from [chart.type](#chart.type).
