+++ js/masters/indicators/indicators-all.src.js
@@ -48,7 +48,7 @@ import '../../Stock/Indicators/VBPIndicator.js';
 import '../../Stock/Indicators/VWAPIndicator.js';
 import '../../Stock/Indicators/WilliamsRIndicator.js';
 import '../../Stock/Indicators/WMAIndicator.js';
-import '../../Stock/Indicators/ZigzagIndicator.js';
+import '../../Stock/Indicators/Zigzag/ZigzagIndicator.js';
 import '../../Stock/Indicators/RegressionIndicators.js';
 import '../../Stock/Indicators/ABands/ABandsIndicator.js';
 import '../../Stock/Indicators/TrendLineIndicator.js';
