+++ ts/Stock/Indicators/ZIGZAG/ZIGZAGPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ZIGZAGIndicator from './ZIGZAGIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class ZIGZAGPoint extends SMAPoint {
+    public series: ZIGZAGIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default ZIGZAGPoint;
